Реализовал сценарии связанные с предложением и пользователем, добавил обработку ошибок, добавил exception-filter

+++ src/modules/offer/offer.controller.ts
@@ -7,10 +7,11 @@ import { AppComponent } from '../../types/app-components.enum.js';
 import { HttpMethods } from '../../types/http-methods.enum.js';
 import { fillDTO } from '../../core/helpers/index.js';
 import OffersRDO from './rdo/offers.rdo.js';
-//import RequestOfferDto from './dto/request-offer.dto.js';
 import OfferRDO from './rdo/offer.rdo.js';
 import CreateOfferDto from './dto/create-offer.dto.js';
-//import UpdateOfferDTO from './dto/update-offer.dto.js';
+import UpdateOfferDTO from './dto/update-offer.dto.js';
+import HTTPError from '../../core/errors/http-error.js';
+import { StatusCodes } from 'http-status-codes';
 
 @injectable()
 export default class OfferController extends Controller {
@@ -35,8 +36,17 @@ export default class OfferController extends Controller {
     this.ok(res, offersResponse);
   };
 
-  public exactOffer = async ({ params }: Request, res: Response) => {
+  public exactOffer = async ({ params }: Request, res: Response): Promise<void> => {
     const offer = await this.offerService.findByOfferId(params.id);
+
+    if (!offer) {
+      throw new HTTPError(
+        StatusCodes.NOT_FOUND,
+        `Offer with id: ${params.id} not found`,
+        'OfferController'
+      );
+    }
+
     const offerResponse = fillDTO(OfferRDO, offer);
     this.ok(res, offerResponse);
   };
@@ -45,17 +55,38 @@ export default class OfferController extends Controller {
     { body }: Request<Record<string, unknown>, Record<string, unknown>, CreateOfferDto>,
     res: Response): Promise<void> => {
     const createdOffer = await this.offerService.create(body);
-    this.created(res, createdOffer);
+    this.created(res, fillDTO(OffersRDO, createdOffer));
   };
 
   public updateOffer = async (
-    req: Request,
+    { params, body }: Request<Record<string, string>, Record<string, unknown>, UpdateOfferDTO>,
     res: Response): Promise<void> => {
-    const updatedOffer = await this.offerService.updateOffer(req.params.id, req.body);
+
+    const offer = this.offerService.findByOfferId(params.id);
+
+    if (!offer) {
+      throw new HTTPError(
+        StatusCodes.NOT_FOUND,
+        `Offer with id: ${params.id} not found`,
+        'OfferController'
+      );
+    }
+
+    const updatedOffer = await this.offerService.updateOffer(params.id, body);
     this.created(res, fillDTO(OfferRDO, updatedOffer));
   };
 
-  public deleteOffer = async ({ params }: Request, res: Response): Promise<void> => {
+  public deleteOffer = async ({ params }: Request<Record<string, string>>, res: Response): Promise<void> => {
+    const offer = this.offerService.findByOfferId(params.id);
+
+    if (!offer) {
+      throw new HTTPError(
+        StatusCodes.NOT_FOUND,
+        `Offer with id: ${params.id} not found`,
+        'OfferController'
+      );
+    }
+
     const deletedOffer = await this.offerService.deleteOffer(params.id);
     this.noContent(res, deletedOffer);
   };
